Moved code to secure_hospital_system
