+++ secure_hospital_system/templates/patient_portal/medical_records.html
@@ -1,4 +1,4 @@
-{% extends 'patient_portal/base.html' %}
+{% extends 'base.html' %}
 {% load render_table from django_tables2 %}
 {% block content %}
